requirements.txt
pystray>=0.19.4
Pillow>=9.0.0
pywin32>=305;platform_system=="Windows"